pyproject.toml
[project]
name = "omnichat-v1"
version = "0.1.0"
description = "AI Assistant with multiple provider support"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask",
    "anthropic",
    "openai",
    "google-cloud-aiplatform",
    "python-dotenv"
]
